+++ WalletWasabi/Helpers/ByteArrayEqualityComparer.cs
@@ -2,13 +2,14 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
+using WalletWasabi.Crypto;
 
 namespace WalletWasabi.Helpers
 {
-	public class ByteArrayEqualityComparer : IEqualityComparer<byte[]>
-	{
-		public bool Equals([AllowNull] byte[] x, [AllowNull] byte[] y) => ByteHelpers.CompareFastUnsafe(x, y);
+    public class ByteArrayEqualityComparer : IEqualityComparer<byte[]>
+    {
+        public bool Equals([AllowNull] byte[] x, [AllowNull] byte[] y) => ByteHelpers.CompareFastUnsafe(x, y);
 
-		public int GetHashCode([DisallowNull] byte[] obj) => HashHelpers.ComputeHashCode(obj);
-	}
+        public int GetHashCode([DisallowNull] byte[] obj) => HashHelpers.ComputeHashCode(obj);
+    }
 }
