+++ WalletWasabi/Helpers/NBitcoinHelpers.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using WalletWasabi.Crypto;
 
 namespace WalletWasabi.Helpers
 {
